restructure files for dashboard components
